refactor(schema-dsl): split field and entity decorators into individual files
Moves each field decorator into its own file under src/fields/ and each
entity (class) decorator into its own file under src/entities/, re-exporting
them all from index.ts. Pure reorganization with no behavioral changes.
