+++ warp-drive-packages/schema-dsl/src/-private/types.ts
@@ -1,0 +1,2 @@
+export type AnyConstructor = abstract new (...args: unknown[]) => unknown;
+export type PrimitiveValue = string | number | boolean | null;
